requirements.txt
absl-py==2.2.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
appdirs==1.4.4
attrs==25.3.0
av==14.3.0
basicsr==1.4.2
beautifulsoup4==4.13.4
blinker==1.9.0
browser-cookie3==0.20.1
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cssselect==1.3.0
ctranslate2==4.6.0
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
defusedxml==0.7.1
demjson3==3.0.6
Deprecated==1.2.18
dropbox==12.0.2
edge-tts==7.2.0
ExifRead==3.2.0
facebook-scraper==0.2.59
facexlib==0.3.0
fake-useragent==2.2.0
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
ghostscript==0.7
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
hf-xet==1.1.7
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
lxml==6.0.0
lxml_html_clean==0.4.2
lz4==4.4.4
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
ml_dtypes==0.5.1
more-itertools==10.7.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.4.3
mutagen==1.47.0
narwhals==1.40.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
onnxruntime==1.21.1
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pikepdf==9.7.0
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.8
playwright==1.54.0
ply==3.11
proglog==0.1.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pygame==2.6.1
PyGithub==2.7.0
PyJWT==2.10.1
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyquery==2.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pyspellchecker==0.8.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.4
requests-html==0.10.0
rpds-py==0.25.1
rsa==4.9.1
schedule==1.2.2
scikit-image==0.25.2
scipy==1.15.2
selenium==4.31.0
sentencepiece==0.2.0
setuptools==80.7.1
shadowcopy==0.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.7
SpeechRecognition==3.14.3
srt==3.5.3
stone==3.3.1
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.20.0a20250517
tenacity==9.1.2
tensorboard-data-server==0.7.2
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.8.0
torchaudio==2.7.0
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.7.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
w3lib==2.3.1
watchdog==6.0.0
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
whisper-ctranslate2==0.5.2
WMI==1.5.1
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.3
yapf==0.43.0
yarl==1.20.0
zipp==3.23.0
